pyproject.toml
[tool.poetry]
name = "shopify-csr-chatbot"
version = "0.1.0"
description = "Initial implementation of a Chatbot for Shopify's CSRs"
authors = ["Bernardo Garcia <bergr7@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
llama-index = "^0.8.29.post1"
openai = "^0.28.0"
llama-hub = "^0.0.31"
weaviate-client = "^3.24.1"
pandas = "^2.1.0"
loguru = "^0.7.2"
ipykernel = "^6.25.2"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
isort = "^5.12.0"
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
